backend/requirements.txt
# FastAPI
fastapi[standard]>=0.118.0
uvicorn[standard]>=0.37.0
pydantic>=2.12.0
pydantic-settings>=2.11.0

# Neo4j
neo4j>=6.0.2

# LangChain
langchain>=0.3.27
langchain-community>=0.3.31
langchain-core>=0.3.78
langchain-ollama>=0.3.10

# Playwright
playwright>=1.55.0

# HTTP Client
httpx>=0.28.1
aiohttp>=3.13.0

# Utilities
python-dotenv>=1.1.1
python-multipart>=0.0.20

# Testing
pytest>=8.4.2
pytest-asyncio>=1.2.0
pytest-cov>=7.0.0

# Development

black>=25.9.0
ruff>=0.14.0
mypy>=1.18.2


firecrawl-py
